+++ src/app.controller.ts
@@ -1,49 +1,14 @@
-import {
-  Body,
-  ClassSerializerInterceptor,
-  Controller,
-  Get,
-  Post,
-  Request,
-  UseGuards,
-  UseInterceptors,
-} from '@nestjs/common';
+import { Controller, Get } from '@nestjs/common';
 import { AppService } from './app.service';
-import { LocalAuthGuard } from './auth/local-auth.guard';
-import { JwtAuthGuard } from './auth/jwt-auth.guard';
-import { AuthService } from './auth/auth.service';
 import { Public } from './common/decorators/public.decorator';
 
 @Controller()
 export class AppController {
-  constructor(
-    private readonly appService: AppService,
-    private readonly authService: AuthService,
-  ) {}
+  constructor(private readonly appService: AppService) {}
 
   @Public()
   @Get()
   getHello(): string {
     return this.appService.getHello();
-  }
-
-  @UseGuards(LocalAuthGuard) // JwtGuard globally set but test this
-  @Post('/auth/login')
-  async login(@Body() body) {
-    return this.authService.login(body);
-  }
-
-  @Public()
-  @Post('/auth/register')
-  async register(@Body() body) {
-    console.log(body);
-    return this.authService.register(body.user);
-  }
-
-  @UseGuards(JwtAuthGuard) // Globally set
-  @UseInterceptors(ClassSerializerInterceptor)
-  @Get('profile')
-  getProfile(@Request() req) {
-    return req.user;
   }
 }
