+++ src/users/users.controller.ts
@@ -1,0 +1,19 @@
+import {
+  ClassSerializerInterceptor,
+  Controller,
+  Get,
+  Request,
+  UseGuards,
+  UseInterceptors,
+} from '@nestjs/common';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
+
+@Controller('users')
+export class UsersController {
+  @UseGuards(JwtAuthGuard) // Globally set
+  @UseInterceptors(ClassSerializerInterceptor)
+  @Get('profile')
+  getProfile(@Request() req) {
+    return req.user;
+  }
+}
