+++ src/users/users.module.ts
@@ -2,12 +2,13 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { Users, UsersSchema } from './entities/users.entity';
+import { UsersController } from './users.controller';
 
 @Module({
   imports: [
     MongooseModule.forFeature([{ name: Users.name, schema: UsersSchema }]),
   ],
-  controllers: [],
+  controllers: [UsersController],
   providers: [UsersService],
   exports: [UsersService],
 })
